+++ src/Client.php
@@ -273,7 +273,7 @@ class Client implements ClientInterface, IteratorAggregate
      * Applies the configured serializer and compression to given value.
      *
      * @param  mixed  $value
-     * @return string
+     * @return mixed
      */
     public function pack($value)
     {
@@ -286,7 +286,7 @@ class Client implements ClientInterface, IteratorAggregate
      * Deserializes and decompresses to given value.
      *
      * @param  mixed  $value
-     * @return string
+     * @return mixed
      */
     public function unpack($value)
     {